refactor: move and merge resolver types to types.resolver
